fix(language-service): include directives in template AST paths
When constructing a path to a template node, the language service
currently expunges all directives ASTs. This commit includes them again,
which simplifies location of symbols where directives are, in fact, the
most specific symbol (like element selectors). This commit will also
help in the implementation of #34564, which relies on determination of
the nearest AST in a path.

One inconsistency this introduces is the location of symbols referencing
components between opening and closing tags in an element. In the past,
querying the definition for both opening and closing tags would result
in a query for the symbol encompassing the entire element.

```text
      v                 cursor
<app-c|omp></app-comp>
~~~~~~~~~~~~~~~~~~~~~~  symbol location
                 v      cursor
<app-comp></app-c|omp>
~~~~~~~~~~~~~~~~~~~~~~  symbol location
```

Now, this behavior only occurs for closing tags; for opening tags, the
queried symbol is the opening tag precisely because that's where the
component selector applies.

```text
      v                 cursor
<app-c|omp></app-comp>
~~~~~~~~~~~             symbol location
                 v      cursor
<app-comp></app-c|omp>
~~~~~~~~~~~~~~~~~~~~~~  symbol location
```